+++ src/deprecate/utils.py
@@ -2,9 +2,32 @@
 
 Copyright (C) 2020-2023 Jiri Borovec <...>
 """
+import inspect
 import warnings
 from contextlib import contextmanager
-from typing import Any, Generator, List, Optional, Type, Union
+from typing import Any, Generator, List, Optional, Type, Union, Callable, Tuple
+
+
+def get_func_arguments_types_defaults(func: Callable) -> List[Tuple[str, Tuple, Any]]:
+    """Parse function arguments, types and default values.
+
+    Args:
+        func: a function to be xeamined
+
+    Returns:
+        sequence of details for each position/keyward argument
+
+    Example:
+        >>> get_func_arguments_types_defaults(get_func_arguments_types_defaults)
+        [('func', typing.Callable, <class 'inspect._empty'>)]
+    """
+    func_default_params = inspect.signature(func).parameters
+    func_arg_type_val = []
+    for arg in func_default_params:
+        arg_type = func_default_params[arg].annotation
+        arg_default = func_default_params[arg].default
+        func_arg_type_val.append((arg, arg_type, arg_default))
+    return func_arg_type_val
 
 
 def _warns_repr(warns: List[warnings.WarningMessage]) -> List[Union[Warning, str]]:
